Document merge_deployment_data()
and make tests more deterministic.
